[FLINK-14041][tests] Refactor LeaderRetrievalServiceHostnameResolutionTest to AddressResolutionTest
